requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2==2.9.9
uvicorn==0.29.0
gunicorn==21.2.0